apps/simple-youtube-downloader/pyproject.toml
[project]
name = "ydl-streamlit"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "hjosugi", email = "h.jo.sugi@azoom.jp" }]
dependencies = [
    "youtube-dl-nightly>=2024.2.19",
    "ffmpeg-python>=0.2.0",
    "streamlit>=1.31.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "ydl_streamlit:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
